Refactoring undate formulas, part ii.
Update formulas for TC relation in DAGs are factored out.

Update formulas for edge relation are same for any topology (factored out).

All imports need to be added at the top level in the same file. See */utests/import_all.vpr for exmaples. This is needed to support using update formulas for different topologies in the same example.